add disable introspection
