+++ graphene/validation/tests/test_disable_introspection.py
@@ -1,0 +1,33 @@
+from graphql import parse, validate
+
+from ...types import Schema, ObjectType, String
+from ..disable_introspection import disable_introspection
+
+
+class Query(ObjectType):
+    name = String(
+        required=True
+    )
+
+
+schema = Schema(query=Query)
+
+
+def run_query(query: str):
+    document = parse(query)
+
+    result = None
+
+    def callback(query_depths):
+        nonlocal result
+        result = query_depths
+
+    errors = validate(
+        schema.graphql_schema,
+        document,
+        rules=(
+            disable_introspection(),
+        ),
+    )
+
+    return errors, result
