--- graphene/validators/__init__.py
@@ -1,6 +1,0 @@
-from .depth_limit_validator import depth_limit_validator
-
-
-__all__ = [
-    "depth_limit_validator"
-]
